cleanup: Moved test functions outside of the GenericSerialFabricator class
