+++ server-javascript/src/serial_fabricator.js
@@ -2,6 +2,7 @@
 
 // Used to communicate with serial devices
 import { SerialPort } from 'serialport';
+import { DEBUG_ENABLED } from './test.js'; /** @todo REPLACE */
 
 /**
  * The anotomy of response extractor type
@@ -31,7 +32,7 @@ import { SerialPort } from 'serialport';
  * It assumes that the fabricator is flashed with {@link https://github.com/MarlinFirmware/Marlin Marlin firmware}
  * @class
  */
-class GenericSerialFabricator {
+export class GenericSerialFabricator {
     /**
      * The dummy instruction to send to the fabricator when there's nothing in the instruction queue
      * @readonly
@@ -345,40 +346,3 @@ class GenericSerialFabricator {
         return await this.sendGCodeInstruction(instruction);
     }
 }
-
-var DEBUG_ENABLED = true; /** @todo BIG TODO: Add ANSI colors to logged outputs to make them easier to read, create a module that will host all of the debug functions, including this function */
-
-class PrusaMK3 extends GenericSerialFabricator {
-    RESPONSE_TIMEOUT = 10000;
-    BOOT_TIME = 5000;
-    INFO_CMD_EXTRACTOR = /FIRMWARE_NAME:(?<firmwareVersion>[^\s]+).+MACHINE_TYPE:(?<machineType>[^\s]+).+/;
-}
-
-/** @todo Add a helper function that determines which specific class to use */
-const fab1 = new PrusaMK3('/dev/ttyACM1');
-const fab2 = new GenericSerialFabricator('/dev/ttyACM0');
-
-// Test sending G-Code
-import * as fs from 'node:fs';
-
-let gcode_file_split = fs.readFileSync('./xyz-cali-cube-micro_MK3.gcode', { encoding: "utf8" }).split('\n');
-
-// Remove empty lines and lines that start with ';'
-let filtered_gcode_file = gcode_file_split.filter(line => line.trim() !== '' && !line.trim().startsWith(';'));
-
-for await (let gcode_line of filtered_gcode_file) {
-    gcode_line = gcode_line.split(';')[0];
-
-    fab1.addGCodeInstructionToQueue({ instruction: gcode_line + '\n' });
-}
-
-gcode_file_split = fs.readFileSync('./xyz-cali-cube-mini_MK4.gcode', { encoding: "utf8" }).split('\n');
-
-// Remove empty lines and lines that start with ';'
-filtered_gcode_file = gcode_file_split.filter(line => line.trim() !== '' && !line.trim().startsWith(';'));
-
-for await (let gcode_line of filtered_gcode_file) {
-    gcode_line = gcode_line.split(';')[0];
-
-    fab2.addGCodeInstructionToQueue({ instruction: gcode_line + '\n' });
-}
